+++ src/Symfony/Component/Form/Extension/Core/Type/TimezoneType.php
@@ -13,7 +13,11 @@ namespace Symfony\Component\Form\Extension\Core\Type;
 
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\ChoiceList\ArrayChoiceList;
+use Symfony\Component\Form\ChoiceList\Loader\CallbackChoiceLoader;
 use Symfony\Component\Form\ChoiceList\Loader\ChoiceLoaderInterface;
+use Symfony\Component\Form\Extension\Core\DataTransformer\DateTimeZoneToStringTransformer;
+use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\OptionsResolver\Options;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
 class TimezoneType extends AbstractType implements ChoiceLoaderInterface
@@ -24,8 +28,20 @@ class TimezoneType extends AbstractType implements ChoiceLoaderInterface
      * The choices are generated from the ICU function \DateTimeZone::listIdentifiers().
      *
      * @var ArrayChoiceList
+     *
+     * @deprecated since version 3.4, to be removed in 4.0
      */
     private $choiceList;
+
+    /**
+     * {@inheritdoc}
+     */
+    public function buildForm(FormBuilderInterface $builder, array $options)
+    {
+        if ('datetimezone' === $options['input']) {
+            $builder->addModelTransformer(new DateTimeZoneToStringTransformer($options['multiple']));
+        }
+    }
 
     /**
      * {@inheritdoc}
@@ -33,9 +49,21 @@ class TimezoneType extends AbstractType implements ChoiceLoaderInterface
     public function configureOptions(OptionsResolver $resolver)
     {
         $resolver->setDefaults(array(
-            'choice_loader' => $this,
+            'choice_loader' => function (Options $options) {
+                $regions = $options['regions'];
+
+                return new CallbackChoiceLoader(function () use ($regions) {
+                    return self::getTimezones($regions);
+                });
+            },
             'choice_translation_domain' => false,
+            'input' => 'string',
+            'regions' => \DateTimeZone::ALL,
         ));
+
+        $resolver->setAllowedValues('input', array('string', 'datetimezone'));
+
+        $resolver->setAllowedTypes('regions', 'int');
     }
 
     /**
@@ -56,21 +84,29 @@ class TimezoneType extends AbstractType implements ChoiceLoaderInterface
 
     /**
      * {@inheritdoc}
+     *
+     * @deprecated since version 3.4, to be removed in 4.0
      */
     public function loadChoiceList($value = null)
     {
+        @trigger_error(sprintf('Method "%s" is deprecated since version 3.4 and will be removed in 4.0.', __METHOD__), E_USER_DEPRECATED);
+
         if (null !== $this->choiceList) {
             return $this->choiceList;
         }
 
-        return $this->choiceList = new ArrayChoiceList(self::getTimezones(), $value);
+        return $this->choiceList = new ArrayChoiceList(self::getTimezones(\DateTimeZone::ALL), $value);
     }
 
     /**
      * {@inheritdoc}
+     *
+     * @deprecated since version 3.4, to be removed in 4.0
      */
     public function loadChoicesForValues(array $values, $value = null)
     {
+        @trigger_error(sprintf('Method "%s" is deprecated since version 3.4 and will be removed in 4.0.', __METHOD__), E_USER_DEPRECATED);
+
         // Optimize
         $values = array_filter($values);
         if (empty($values)) {
@@ -87,9 +123,13 @@ class TimezoneType extends AbstractType implements ChoiceLoaderInterface
 
     /**
      * {@inheritdoc}
+     *
+     * @deprecated since version 3.4, to be removed in 4.0
      */
     public function loadValuesForChoices(array $choices, $value = null)
     {
+        @trigger_error(sprintf('Method "%s" is deprecated since version 3.4 and will be removed in 4.0.', __METHOD__), E_USER_DEPRECATED);
+
         // Optimize
         $choices = array_filter($choices);
         if (empty($choices)) {
@@ -107,13 +147,15 @@ class TimezoneType extends AbstractType implements ChoiceLoaderInterface
     /**
      * Returns a normalized array of timezone choices.
      *
+     * @param int $regions
+     *
      * @return array The timezone choices
      */
-    private static function getTimezones()
+    private static function getTimezones($regions)
     {
         $timezones = array();
 
-        foreach (\DateTimeZone::listIdentifiers() as $timezone) {
+        foreach (\DateTimeZone::listIdentifiers($regions) as $timezone) {
             $parts = explode('/', $timezone);
 
             if (count($parts) > 2) {
@@ -130,6 +172,6 @@ class TimezoneType extends AbstractType implements ChoiceLoaderInterface
             $timezones[$region][str_replace('_', ' ', $name)] = $timezone;
         }
 
-        return $timezones;
+        return 1 === count($timezones) ? reset($timezones) : $timezones;
     }
 }