+++ src/Symfony/Component/Form/CHANGELOG.md
@@ -23,6 +23,8 @@ CHANGELOG
 -----
 
  * added `DebugCommand`
+ * deprecated `ChoiceLoaderInterface` implementation in `TimezoneType`
+ * added options "input" and "regions" to `TimezoneType`
 
 3.3.0
 -----